+++ t.App/Controls/PlayedCardsContentView.xaml
@@ -5,7 +5,7 @@
              xmlns:converter="clr-namespace:t.App.Controls.Converter;assembly=t.App.Controls"
              x:Class="t.App.Controls.PlayedCardsContentView"
              xmlns:localcontrols="clr-namespace:t.App.Controls;assembly=t.App"
-             xmlns:localconverter="clr-namespace:t.App">
+             xmlns:localconverter="clr-namespace:t.App.Converter">
     <ContentView.Resources>
 
         <localconverter:PlayerColumnDefinitionConverter x:Key="PlayerColumnDefinitionConverter" />
