+++ t.App/Controls/SelectedCardPlayerContainer.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:controls="clr-namespace:t.App.Controls;assembly=t.App.Controls"
              xmlns:converter="clr-namespace:t.App.Controls.Converter;assembly=t.App.Controls"
-             xmlns:localconverter="clr-namespace:t.App"
+             xmlns:localconverter="clr-namespace:t.App.Converter"
              x:Class="t.App.Controls.SelectedCardPlayerContainer">
     <ContentView.Resources>
