+++ t.App/View/GamePage.xaml
@@ -8,6 +8,7 @@
              xmlns:animations="clr-namespace:t.App.Controls.Animations;assembly=t.App.Controls"
              xmlns:converter="clr-namespace:t.App.Controls.Converter;assembly=t.App.Controls"
              xmlns:localcontrols="clr-namespace:t.App.Controls;assembly=t.App"
+             xmlns:localconverter="clr-namespace:t.App.Converter"
              Title="{Binding Path=GameClientViewModel.Title}"
              NavigationPage.HasNavigationBar="false"
              x:Class="t.App.View.GamePage">
@@ -17,7 +18,7 @@
 
         <converter:ValuesComparerMultiValueConverter x:Key="ValuesComparerMultiValueConverter" />
 
-
+        <localconverter:ToItemsLayoutOrientationConverter x:Key="ToItemsLayoutOrientationConverter" />
 
 
         <Style x:Key="cardView" TargetType="{x:Type controls:CardView}">
@@ -111,9 +112,9 @@
                 <Grid Grid.Row="0" Grid.ColumnSpan="{OnPlatform Android=2,Default=0}" Grid.RowSpan="{OnPlatform Android=0,Default=2}">
                     <CollectionView ItemsSource="{Binding Players}" SelectionMode="None" HorizontalOptions="CenterAndExpand">
                         <CollectionView.ItemsLayout>
-                            <LinearItemsLayout Orientation="Vertical">
-                            </LinearItemsLayout>
-                            <!--<LinearItemsLayout Orientation="{OnPlatform Android='Horizontal',Default=}" />-->
+                            <LinearItemsLayout Orientation="Vertical" />
+                            <!--<LinearItemsLayout Orientation="{OnPlatform Android='Horizontal',Default=Vertical}" />-->
+                            
                             <!--<OnPlatform x:TypeArguments="LinearItemsLayout">
                                 <On Platform="Default">
                                     <LinearItemsLayout Orientation="Vertical" />
